Increase cache limit from 1 day -> 3 days.
Why?

The system when you view a profile that is over 1 day old will auto trigger a refresh. This is great, but with almost 20,000 gamertags
80% of them are always ready for a refresh. So these damn spiders trigger all of them. 3 days is actually just delaying the problem
and not actually fixing it, but lets the backend recacher and other things go first, but it'll keep the site active and fast for those you utilize it.